Fix critical bugs, add build system, rebalance levels for itch.io launch
Bug fixes:
- game.py: remove duplicate clock.tick() halving FPS to 30
- game.py: fix check_game_over comparing Level object to integer
- game.py: fix update_ui() corrupting frame timing via clock.tick()
- game_state_manager.py: fix KeyError crash on defeat screen (fresh game)
- game_state_manager.py: use player.unlocked_levels directly (authoritative)
- level_manager.py: fix call to non-existent start_next_level()
- level.py: fix modify-list-while-iterating bug in wave cleanup
- bullet.py: remove module-level example code crashing on import
- enemy_wave.py: set is_active=True when wave starts (was never activating)
- enemy_manager.py: remove duplicate enemy spawning (already in game.py)
- tower_manager.py: remove per-frame attack_speed accumulation from skills
- tower.py: replace Ellipsis placeholders with None/[]
- tower_types.py: remove invalid extra kwargs in FrostTower/GoldBoostTower
- collision_manager.py: dokillb=False, let projectile handle own lifecycle
- helpers.py: use resource_path() for frozen-exe compatibility
- entity.py: draw at rect.x/y not x/y (correct sprite position)
- wave_panel.py: fix set_relative_position arg type + bounds check
- tower_selection_panel.py: int() cast floats for pixel dimensions
- options_screen.py: remove bad module-level audio_manager import
- skills_screen.py: skill buttons now tracked in skill_buttons list
- game_data_screen.py: safe fallback for empty/missing unlocked_levels
- asset_manager.py: remove reference to non-existent TOWER_IMAGE_PATH
- Player.py: remove duplicate _update_ui method definition
- level_completion.py: replay passes current level index; remove double image load

Performance:
- game_board.py: create_path_layout once per draw (was 384x per frame)

Balance:
- config.py: PLAYER_GOLD 10000 -> 300 (economy reset)
- LevelsAll.json: levels 1/2/3/5/7 rebalanced (early game survivable)

Build system:
- requirements.txt: pygame==2.5.2, pygame-gui==0.5.4
- build.spec: PyInstaller one-folder Windows build
- build.bat: full build + zip pipeline for itch.io
- src/utils/resource_path.py: frozen/dev path resolution
- src/main.py: os.chdir() to bundle root for frozen asset paths
- setup.py: project metadata

+++ src/board/tower_selection_panel.py
@@ -7,9 +7,9 @@ class TowerSelectionPanel:
     def __init__(self, screen, tower_manager):
         self.screen = screen
         self.tower_manager = tower_manager
-        self.panel_height = screen.get_height() * 0.20  # Adjust as needed
+        self.panel_height = int(screen.get_height() * 0.20)  # Adjust as needed
         self.panel_y = screen.get_height() - self.panel_height
-        self.icon_size = screen.get_width() * 0.07  # Adjust based on your icon sizes
+        self.icon_size = int(screen.get_width() * 0.07)  # Adjust based on your icon sizes
         self.font = pygame.font.Font(None, UI_FONT_SIZE)  # Adjust font as needed
         self.icons = {}  # To store preloaded and scaled images
         self.icon_padding = 5  # 5 pixels of padding
